requirements.txt
requests==2.27.1
environs==9.5.0
transliterate==1.10.2
python-telegram-bot==13.11
redis==4.3.4
geopy==2.2.0
